+++ sandbox/settings.py
@@ -126,9 +126,12 @@ class Base(Configuration):
     ]
 
     AUTHENTICATION_BACKENDS = [
-        "fun_lti_provider.default.backend.AuthBackend",
+        "ashley.auth.backend.LTIBackend",
         "django.contrib.auth.backends.ModelBackend",
     ]
+
+    LTI_LAUNCH_SUCCESS_HANDLER = "ashley.auth.handlers.success"
+    LTI_LAUNCH_FAILURE_HANDLER = "ashley.auth.handlers.failure"
 
     # Django applications from the highest priority to the lowest
     INSTALLED_APPS = [